Rename try_sink.rs to sink.rs
